Organizing, move graph data service to services
