+++ src/app/_services/index.ts
@@ -1,6 +1,7 @@
 export * from './auth.service';
 export * from './CID.service';
 export * from './google-analytics.service';
+export * from './graph-data.service';
 export * from './loader-state.service';
 export * from './login-state.service';
 export * from './search-results.service';
